+++ packages/codec/lib/decode.ts
@@ -4,7 +4,7 @@ const debug = debugModule("codec:decode");
 import * as Abi from "@truffle/codec/abi";
 import * as Format from "@truffle/codec/format";
 import * as Pointer from "@truffle/codec/pointer";
-import * as Elementary from "@truffle/codec/elementary";
+import * as Basic from "@truffle/codec/basic";
 import * as Evm from "@truffle/codec/evm";
 import { DecoderRequest, DecoderOptions } from "@truffle/codec/types";
 import * as Memory from "@truffle/codec/memory";
@@ -32,7 +32,7 @@ export default function* decode(
       return yield* Stack.Decode.decodeLiteral(dataType, pointer, info);
 
     case "definition":
-      return yield* Elementary.Decode.decodeConstant(dataType, pointer, info);
+      return yield* Basic.Decode.decodeConstant(dataType, pointer, info);
 
     case "special":
       return yield* Special.Decode.decodeSpecial(dataType, pointer, info);
