+++ packages/codec/lib/stack/decode/index.ts
@@ -5,7 +5,7 @@ import * as Abi from "@truffle/codec/abi";
 import * as Conversion from "@truffle/codec/conversion";
 import * as Format from "@truffle/codec/format";
 import read from "@truffle/codec/read";
-import * as Elementary from "@truffle/codec/elementary";
+import * as Basic from "@truffle/codec/basic";
 import * as Memory from "@truffle/codec/memory";
 import * as Storage from "@truffle/codec/storage";
 import * as Pointer from "@truffle/codec/pointer";
@@ -68,7 +68,7 @@ export function* decodeLiteral(
         //next: do we have a calldata pointer?
 
         //if it's a string or bytes, we will interpret the pointer ourself and skip
-        //straight to decodeElementary.  this is to allow us to correctly handle the
+        //straight to decodeBasic.  this is to allow us to correctly handle the
         //case of msg.data used as a mapping key.
         if (dataType.typeClass === "bytes" || dataType.typeClass === "string") {
           let startAsBN = Conversion.toBN(
@@ -116,11 +116,7 @@ export function* decodeLiteral(
             start,
             length
           };
-          return yield* Elementary.Decode.decodeElementary(
-            dataType,
-            newPointer,
-            info
-          );
+          return yield* Basic.Decode.decodeBasic(dataType, newPointer, info);
         }
 
         //otherwise, is it a dynamic array?
@@ -155,13 +151,13 @@ export function* decodeLiteral(
   }
 
   //next: do we have an external function?  these work differently on the stack
-  //than elsewhere, so we can't just pass it on to decodeElementary.
+  //than elsewhere, so we can't just pass it on to decodeBasic.
   if (dataType.typeClass === "function" && dataType.visibility === "external") {
     let address = pointer.literal.slice(0, Evm.Utils.WORD_SIZE);
     let selectorWord = pointer.literal.slice(-Evm.Utils.WORD_SIZE);
     if (
-      !Elementary.Decode.checkPaddingLeft(address, Evm.Utils.ADDRESS_SIZE) ||
-      !Elementary.Decode.checkPaddingLeft(selectorWord, Evm.Utils.SELECTOR_SIZE)
+      !Basic.Decode.checkPaddingLeft(address, Evm.Utils.ADDRESS_SIZE) ||
+      !Basic.Decode.checkPaddingLeft(selectorWord, Evm.Utils.SELECTOR_SIZE)
     ) {
       return {
         type: dataType,
@@ -177,19 +173,15 @@ export function* decodeLiteral(
     return {
       type: dataType,
       kind: "value" as const,
-      value: yield* Elementary.Decode.decodeExternalFunction(
-        address,
-        selector,
-        info
-      )
+      value: yield* Basic.Decode.decodeExternalFunction(address, selector, info)
     };
   }
 
-  //finally, if none of the above hold, we can just dispatch to decodeElementary.
+  //finally, if none of the above hold, we can just dispatch to decodeBasic.
   //however, note that because we're on the stack, we use the permissive padding
   //option so that errors won't result due to values with bad padding
   //(of numeric or bytesN type, anyway)
-  return yield* Elementary.Decode.decodeElementary(dataType, pointer, info, {
+  return yield* Basic.Decode.decodeBasic(dataType, pointer, info, {
     permissivePadding: true
   });
 }
